Collapse action creators into action files
